Ex6/requirements.txt
functions_framework==3.7.0
protobuf>=4.21.6,<5.0.0
google-cloud-storage==2.18.0
cloudevents==1.11.0
google-cloud-bigquery==3.25.0